requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep learning
torch>=1.12.0
torchvision>=0.13.0

# Time series analysis
statsmodels>=0.13.0
pmdarima>=2.0.0
prophet>=1.1.0
tslearn>=0.5.0
darts>=0.20.0
sktime>=0.15.0

# Visualization
plotly>=5.0.0
streamlit>=1.20.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional: Jupyter support
jupyter>=1.0.0
ipywidgets>=8.0.0
